chore: finish up context menu code
