+++ TransparentCloudServerProxy.Client/ViewModels/Pages/RemoteServersViewModel.cs
@@ -1,12 +1,7 @@
 ﻿using System.Collections.ObjectModel;
-using System.Reactive;
-using System.Threading.Tasks;
 
 using DynamicData;
 
-using Microsoft.AspNetCore.Hosting.Server;
-
-using ReactiveUI;
 using ReactiveUI.Fody.Helpers;
 
 using TransparentCloudServerProxy.Client.Services.Api;
@@ -15,26 +10,17 @@ using TransparentCloudServerProxy.Client.ViewModels.Controls;
 
 namespace TransparentCloudServerProxy.Client.ViewModels.Pages {
     public class RemoteServersViewModel : ViewModel {
-        private IProxyServerService _proxyServerService { get; set; }
-
         public ObservableCollection<ProxyServer> Servers { get; } = new();
 
         public ObservableCollection<ProxyDataGridViewModel> ServerDataGrids { get; } = new();
 
-
-
         [Reactive]
         public bool ApplyChangesVisible { get; set; }
 
-        public ReactiveCommand<Unit, Unit> DeleteCommand { get; }
-        public ReactiveCommand<Unit, Unit> ResetChanges { get; }
 
-
+        private readonly IProxyServerService _proxyServerService;
         public RemoteServersViewModel(IProxyServerService proxyServerService) {
             _proxyServerService = proxyServerService;
-
-            DeleteCommand = ReactiveCommand.CreateFromTask(DeleteCommandAsync);
-            ResetChanges = ReactiveCommand.CreateFromTask(ResetChangesAsync);
 
             SyncServers();
         }
@@ -48,13 +34,6 @@ namespace TransparentCloudServerProxy.Client.ViewModels.Pages {
             foreach (var proxyServer in Servers) {
                 ServerDataGrids.Add(new ProxyDataGridViewModel(proxyServer));
             }
-        }
-
-        public async Task DeleteCommandAsync() {
-
-        }
-
-        public async Task ResetChangesAsync() {
         }
     }
 }
